refactor: moved common helper functions to dsa_api_handler
